requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
Pillow
torch
torchvision
torchxrayvision
openai
streamlit
requests
